// libs/enums/product.enum.ts
export enum ProductType {
	PET = 'PET',
	FOOD = 'FOOD',
	TOY = 'TOY',
	ACCESSORY = 'ACCESSORY',
}

export enum ProductSpecies {
	DOG = 'DOG',
	CAT = 'CAT',
	BIRD = 'BIRD',
	FISH = 'FISH',
}

export enum ProductGender {
	MALE = 'MALE',
	FEMALE = 'FEMALE',
}

export enum ProductStatus {
	ACTIVE = 'ACTIVE',
	SOLD = 'SOLD',
	DELETE = 'DELETE',
}

export enum ProductLocation {
	SEOUL = 'SEOUL',
	BUSAN = 'BUSAN',
	INCHEON = 'INCHEON',
	DAEGU = 'DAEGU',
	GYEONGJU = 'GYEONGJU',
	GWANGJU = 'GWANGJU',
	CHONJU = 'CHONJU',
	DAEJON = 'DAEJON',
	JEJU = 'JEJU',
}